RegisterDeformation: build a "real" Deformation type
This is rebased on FixedSizeArrays, and the image size (really,
knot positions) is now stored along with u in a deformation
type. This allows one to implement composition, which was always
a major pain back when we used a plain u array, and seems likely
to make many things easier.