Add CachedInterpolations for somewhat faster optimization
This type snips out a 3x3x3 block of the current aperture's mismatch data, centered on the locus of interpolation. Since it packs the data into a tighter space, one expects fewer page faults/cache misses (and perhaps fewer trips to the disk, for swap/mmap) when evaluating the data penalty.