Limit the bias-correction data to the same z-slice
It turns out that incorporating information from the 3rd dimension was messing up the correction, perhaps because of the lower resolution along the z-axis. Restricting to shifts in just the first two dimensions makes it look much better, at least as good as another scheme I experimented with (performing a local quadratic fit).